move migrations to db
